backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
google-generativeai==0.8.4
chromadb==0.5.23
pypdf2==3.0.1
python-docx==1.1.2
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0